renaming files, stabilizing tests
